Configure waitForWeb3
